+++ DiscordBot/BackendRelated/IChannel.cs
@@ -1,0 +1,7 @@
+namespace DiscordBot.BackendRelated
+{
+    public interface IChannel<T>
+    {
+        T GetChannel();
+    }
+}
