+++ DiscordBot/BackendRelated/IMessage.cs
@@ -1,0 +1,8 @@
+namespace DiscordBot.BackendRelated
+{
+    public interface IMessage
+    {
+        System.Type AuthorType { get; set; }
+        System.Type ChannelType { get; set; }
+    }
+}
