+++ DiscordBot/BackendRelated/IUser.cs
@@ -1,0 +1,7 @@
+namespace DiscordBot.BackendRelated
+{
+    public interface IUser<T>
+    {
+        T GetUser();
+    }
+}
