+++ DiscordBot/BackendRelated/Twitter/TwitterBackend.cs
@@ -1,0 +1,9 @@
+using Tweetinvi;
+
+namespace DiscordBot.BackendRelated.Twitter
+{
+    public class TwitterBackend
+    {
+        private TwitterClient _client;
+    }
+}
