+++ DiscordBot/BackendRelated/Twitter/TwitterContext.cs
@@ -1,0 +1,9 @@
+using Tweetinvi;
+
+namespace DiscordBot.BackendRelated.Twitter
+{
+    public class TwitterContext
+    {
+        
+    }
+}
