--- DiscordBot/Backends/TicTacToeManager.cs
@@ -1,47 +1,0 @@
-using System;
-using System.Collections.Generic;
-using DSharpPlus.Entities;
-using TicTacToe;
-
-namespace DiscordBot.Backends
-{
-    public class TicTacToeManager<T>
-    {
-        private Dictionary<Tuple<DiscordUser, DiscordChannel>, TicTacToeGame<T>> Players { get; set; }
-
-        public TicTacToeManager() =>
-            Players = new Dictionary<Tuple<DiscordUser, DiscordChannel>, TicTacToeGame<T>>();
-
-        public bool AddGame(IContext msg, Dictionary<T, SquareState> translationDict)
-        {
-            var gameExists = GameExists(GetGame(msg));
-            
-            if (!gameExists)
-                Players.Add(TupleFromMessage(msg), new TicTacToeGame<T>(translationDict));
-
-            return !gameExists;
-        }
-
-        public bool RemoveGame(IContext msg)
-        {
-            var gameExists = GameExists(GetGame(msg));
-            
-            if (gameExists)
-                Players.Remove(TupleFromMessage(msg));
-
-            return gameExists;
-        }
-        
-        public TicTacToeGame<T> GetGame(IContext msg)
-        {
-            Players.TryGetValue(TupleFromMessage(msg), out var game);
-            return game;
-        }
-
-        public bool GameExists(TicTacToeGame<T> game) =>
-            game != null;
-        
-        private static Tuple<DiscordUser, DiscordChannel> TupleFromMessage(IContext msg) =>
-            new Tuple<DiscordUser, DiscordChannel>(msg.GetSender(), msg.GetChannel());
-    }
-}
