+++ DiscordBot/CommandRelated/Commands/PingCommand.cs
@@ -1,0 +1,13 @@
+﻿using System.Threading.Tasks;
+using DiscordBot.BackendRelated;
+
+namespace DiscordBot.CommandRelated.Commands
+{
+    class PingCommand: ICommand
+    {
+        public async Task Run(IContext e)
+        {
+            await e.Respond("pong!");
+        }
+    }
+}
