+++ DiscordBot/CommandRelated/ICommandProvider.cs
@@ -1,0 +1,12 @@
+﻿using System.Collections.Generic;
+using DiscordBot.CommandRelated.Commands;
+
+namespace DiscordBot.CommandRelated
+{
+    interface ICommandProvider
+    {
+        Dictionary<string[], ICommand> Commands { get; set; }
+        
+        ICommand ProvideCommand(string content);
+    }
+}
