--- DiscordBot/Commands/ICommand.cs
@@ -1,16 +1,0 @@
-﻿using System.Threading.Tasks;
-using DiscordBot.Backends;
-
-namespace DiscordBot.Commands
-{
-    // Commands that only send a message in response
-    public interface ICommand
-    {
-        Task Run(IContext e);
-
-        Task OnMessage(IContext e)
-        {
-            return Task.CompletedTask;
-        }
-    }
-}
