--- DiscordBot/Config/Config.cs
@@ -1,8 +1,0 @@
-﻿namespace DiscordBot.Config
-{
-    public class Config
-    {
-        public string Token { get; set; }
-        public string Prefix { get; set; }
-    }
-}
