+++ DiscordBot/Config/DiscordConfig.cs
@@ -1,0 +1,8 @@
+﻿namespace DiscordBot.Config
+{
+    public class DiscordConfig: IConfig
+    {
+        public string Token { get; set; }
+        public string Prefix { get; set; }
+    }
+}
