+++ DiscordBot/Config/IConfig.cs
@@ -1,0 +1,7 @@
+namespace DiscordBot.Config
+{
+    public interface IConfig
+    { 
+        string Prefix { get; set; }
+    }
+}
