+++ DiscordBot/Config/IConfigGetter.cs
@@ -2,6 +2,6 @@ namespace DiscordBot.Config
 {
     public interface IConfigGetter
     {
-        Config GetConfig();
+        IConfig GetConfig();
     }
 }
