+++ DiscordBot/Config/JsonConfigGetter.cs
@@ -1,4 +1,5 @@
-﻿using System.IO;
+﻿using System;
+using System.IO;
 using Newtonsoft.Json;
 
 namespace DiscordBot.Config
@@ -6,23 +7,28 @@ namespace DiscordBot.Config
     class JsonConfigGetter : IConfigGetter
     {
         private string Path { get; set; }
-        public JsonConfigGetter(string path)
+        private System.Type ConfigType { get; set; }
+        public JsonConfigGetter(string path, System.Type configType)
         {
             Path = path;
+            ConfigType = configType;
         }
         
         /// <summary>
         /// Returns the configurations for the bot
         /// </summary>
-        public Config GetConfig()
+        public IConfig GetConfig()
         {
+            using StreamReader r = new StreamReader(Path);
+            string json = r.ReadToEnd();
             
-            using (StreamReader r = new StreamReader(Path))
-            {
-                string json = r.ReadToEnd();
-                var config = JsonConvert.DeserializeObject<Config>(json);
-                return config;
-            }
+            IConfig config = null;
+            if (ConfigType == typeof(DiscordConfig))
+                config = JsonConvert.DeserializeObject<DiscordConfig>(json);
+            else if (ConfigType == typeof(TwitterConfig))
+                config = JsonConvert.DeserializeObject<TwitterConfig>(json);
+            
+            return config;
         }
