+++ DiscordBot/TicTacToeManager.cs
@@ -1,6 +1,6 @@
 using System;
 using System.Collections.Generic;
-using DiscordBot.Backends;
+using DiscordBot.BackendRelated;
 using DSharpPlus;
 using DSharpPlus.Entities;
 using TicTacToe;
