+++ TicTacToe/TicTacToe.csproj
@@ -4,4 +4,8 @@
         <TargetFramework>netcoreapp3.1</TargetFramework>
     </PropertyGroup>
 
+    <ItemGroup>
+      <PackageReference Include="TweetinviAPI" Version="5.0.3" />
+    </ItemGroup>
+
 </Project>
